fix: :bug: prevents states of child card being destroyed by parent
