+++ packages/cfd/src/Containers/props.types.ts
@@ -241,31 +241,43 @@ type TOpenAccountTransferMeta = {
     type?: string;
 };
 
-export type TJurisdictionModalProps = {
+export type TJurisdictionModalContentWrapperProps = {
+    openPasswordModal: (account_type: TOpenAccountTransferMeta) => void;
+    context: RootStore;
     account_type: {
         type: string;
         category: string;
     };
     account_status: GetAccountStatus;
-    context: RootStore;
-    disableApp: () => void;
-    enableApp: () => void;
     is_jurisdiction_modal_visible: boolean;
     is_virtual: boolean;
     jurisdiction_selected_shortcode: string;
-    openPasswordModal: (account_type: TOpenAccountTransferMeta) => void;
+    real_financial_accounts_existing_data: TExistingData;
+    real_synthetic_accounts_existing_data: TExistingData;
+    trading_platform_available_accounts: TTradingPlatformAvailableAccount[];
+    toggleJurisdictionModal: () => void;
     setJurisdictionSelectedShortcode: (shortcode: string) => void;
     should_restrict_bvi_account_creation: boolean;
     should_restrict_vanuatu_account_creation: boolean;
     show_eu_related_content: boolean;
-    trading_platform_available_accounts: TTradingPlatformAvailableAccount[];
-    fetchAccountSettings: () => void;
-    toggleJurisdictionModal: () => void;
     toggleCFDVerificationModal: () => void;
-    real_synthetic_accounts_existing_data: TExistingData;
-    real_financial_accounts_existing_data: TExistingData;
     updateMT5Status: () => void;
+    fetchAccountSettings: () => void;
     has_submitted_cfd_personal_details: boolean;
+};
+
+export type TJurisdictionModalProps = {
+    openPasswordModal: (account_type: TOpenAccountTransferMeta) => void;
+    context: RootStore;
+    account_type: {
+        type: string;
+        category: string;
+    };
+    disableApp: () => void;
+    enableApp: () => void;
+    is_jurisdiction_modal_visible: boolean;
+    show_eu_related_content: boolean;
+    toggleJurisdictionModal: () => void;
 };
 
 export type TJurisdictionModalContentProps = {
